fix: only replace path when on /trees route
Reason: When on the /trees/[id] path we still have the map rendered. So when we e.g. flip the screen orientation, a MapLibre GL event is triggered that would try to update the viewport and therefore the URL. This results in a router error because the tree [id] is not present in the replace action. By restricting the replace to only happen on /trees we prevent that error.