feat: display nowcast from yesterday on map
The most recent nowcast value is actually always the value fron one day ago. Therefore we need to adjust the outdated filter. Also, while the type_id for average remains 4, the naming has not been changed yet because because it would require some refactoring, including in the vector tiles generator.